requirements.txt
autopep8==1.5.7
numpy==1.21.2
pandas==1.3.3
pycodestyle==2.7.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
selenium==3.141.0
six==1.16.0
toml==0.10.2
urllib3==1.26.7
